Move completion into own provider class
